massive cleanup of bot shutdown, fix #729, fix #993

+++ helpers/workers.py
@@ -15,10 +15,10 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 import re
+import signal
 import threading
 import multiprocessing
 import concurrent.futures
-import logging
 from collections import namedtuple
 from threading import Timer
 from . import backtrace, babble, control
@@ -35,7 +35,7 @@ class Workers():
 
     def __init__(self, handler):
         with worker_lock:
-            self.pool = multiprocessing.Pool()
+            self.pool = multiprocessing.Pool(initializer=self.pool_init)
             self.events = {}
         with executor_lock:
             self.executor = concurrent.futures.ThreadPoolExecutor(4)
@@ -46,6 +46,10 @@ class Workers():
             handler.send(target, handler.config['core']['nick'], msg, 'privmsg')
         self.defer(3600, False, self.handle_pending, handler, send)
         self.defer(3600, False, self.check_babble, handler, send)
+
+    def pool_init(_):
+        """We ignore Ctrl-C in the poll workers, so that we can clean things up properly."""
+        signal.signal(signal.SIGINT, signal.SIG_IGN)
 
     def start_thread(self, func, *args, **kwargs):
         with executor_lock:
@@ -60,7 +64,7 @@ class Workers():
         with worker_lock:
             self.pool.terminate()
             self.pool.join()
-            self.pool = multiprocessing.Pool()
+            self.pool = multiprocessing.Pool(initializer=self.pool_init)
 
     def run_action(self, func, args):
         try:
@@ -87,29 +91,16 @@ class Workers():
                 self.events[eventid].event.function(**self.events[eventid].event.kwargs)
             del self.events[eventid]
 
-    def stop_workers(self):
-        """ Cleanly stop workers and deferred events """
+    def stop_workers(self, clean):
+        """ Stop workers and deferred events """
         with executor_lock:
-            self.executor.shutdown(True)
+            self.executor.shutdown(clean)
             del self.executor
         with worker_lock:
-            self.pool.close()
-            self.pool.join()
-            del self.pool
-            for x in self.events.values():
-                x.event.cancel()
-            self.events.clear()
-
-    def kill_workers(self):
-        """ Forcibly kill all worker threads """
-        with executor_lock:
-            logging.info("Forcibly shutting down executor")
-            self.executor.shutdown(False)
-            del self.executor
-        with worker_lock:
-            logging.info("Killing pool")
-            self.pool.terminate()
-            logging.info("Joining pool")
+            if clean:
+                self.pool.close()
+            else:
+                self.pool.terminate()
             self.pool.join()
             del self.pool
             for x in self.events.values():
